Seed experiment auth state on server
